+++ docs/app/components/Header.vue
@@ -32,12 +32,12 @@ const navigation = inject<Ref<NavItem[]>>('navigation')
         <UContentSearchButton />
       </UTooltip>
 
-      <!-- <UColorModeButton /> -->
+      <ColorModeButton />
 
       <UButton
         color="gray"
         variant="ghost"
-        to="https://github.com/nuxt/ui"
+        to="https://github.com/nuxt/ui/tree/v3"
         target="_blank"
         icon="i-simple-icons-github"
         aria-label="GitHub"